Move store.js and initalState.js into store directory
