+++ client/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import {Provider} from 'react-redux';
 import {Router, browserHistory} from 'react-router';
-import store from './store.js';
+import store from './store';
 import routes from './routes.jsx';
 import MessageBoxes from './components/MessageBoxes.jsx';
 
@@ -10,5 +10,5 @@ ReactDOM.render(
   <Provider store={store}>
     <Router routes={routes} />
   </Provider>,
-  document.getElementById("app")
+  document.getElementById('app')
 );
