+++ client/store/index.js
@@ -1,6 +1,7 @@
 import { createStore, applyMiddleware } from 'redux';
+import rootReducer from './../reducers/rootReducer.js';
+import initialState from './initialState.js';
 import thunkMiddleware from 'redux-thunk';
-import initialState from './initialState.js'
 
 export default createStore(
   rootReducer,
